Fix #1820: Update status endpoint for crypto4
