+++ powerauth-admin/src/main/java/com/wultra/security/app/admin/controller/ActivationController.java
@@ -16,6 +16,7 @@
 
 package com.wultra.security.app.admin.controller;
 
+import com.wultra.security.powerauth.client.model.response.v3.GetActivationStatusResponse;
 import com.wultra.security.powerauth.client.v3.PowerAuthClient;
 import com.wultra.security.powerauth.client.model.entity.Activation;
 import com.wultra.security.powerauth.client.model.entity.ActivationHistoryItem;
