+++ powerauth-client-model/src/main/java/com/wultra/security/powerauth/client/model/response/v3/GetActivationStatusResponse.java
@@ -1,6 +1,6 @@
 /*
  * PowerAuth Server and related software components
- * Copyright (C) 2021 Wultra s.r.o.
+ * Copyright (C) 2025 Wultra s.r.o.
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Affero General Public License as published
@@ -14,9 +14,10 @@
  *
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ *
  */
 
-package com.wultra.security.powerauth.client.model.response;
+package com.wultra.security.powerauth.client.model.response.v3;
 
 import com.wultra.security.powerauth.client.model.enumeration.ActivationOtpValidation;
 import com.wultra.security.powerauth.client.model.enumeration.ActivationStatus;
@@ -30,7 +31,7 @@ import java.util.Date;
 import java.util.List;
 
 /**
- * Model class representing response with activation status.
+ * Model class representing response with activation status (V3).
  *
  * @author Petr Dvorak, petr@wultra.com
  */
