+++ powerauth-java-server/src/main/java/com/wultra/security/powerauth/app/server/controller/api/v3/ActivationController.java
@@ -21,14 +21,17 @@ package com.wultra.security.powerauth.app.server.controller.api.v3;
 
 import com.wultra.security.powerauth.app.server.service.behavior.tasks.ActivationInitServiceBehavior;
 import com.wultra.security.powerauth.app.server.service.behavior.tasks.v3.ActivationCreateServiceBehavior;
+import com.wultra.security.powerauth.app.server.service.behavior.tasks.v3.ActivationStatusServiceBehavior;
 import com.wultra.security.powerauth.client.model.request.*;
 import com.wultra.security.powerauth.client.model.request.v3.CreateActivationRequest;
+import com.wultra.security.powerauth.client.model.request.v3.GetActivationStatusRequest;
 import com.wultra.security.powerauth.client.model.request.v3.PrepareActivationRequest;
 import com.wultra.security.powerauth.client.model.response.*;
 import com.wultra.core.rest.model.base.request.ObjectRequest;
 import com.wultra.core.rest.model.base.response.ObjectResponse;
 import com.wultra.security.powerauth.app.server.service.behavior.tasks.ActivationServiceBehavior;
 import com.wultra.security.powerauth.client.model.response.v3.CreateActivationResponse;
+import com.wultra.security.powerauth.client.model.response.v3.GetActivationStatusResponse;
 import com.wultra.security.powerauth.client.model.response.v3.PrepareActivationResponse;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
@@ -54,8 +57,9 @@ import org.springframework.web.bind.annotation.RestController;
 public class ActivationController {
 
     private final ActivationServiceBehavior activationServiceBehavior;
-    private final ActivationCreateServiceBehavior activationServiceBehaviorV3;
+    private final ActivationCreateServiceBehavior activationCreateServiceBehavior;
     private final ActivationInitServiceBehavior activationInitServiceBehavior;
+    private final ActivationStatusServiceBehavior activationStatusServiceBehavior;
 
     /**
      * Init activation.
@@ -87,7 +91,7 @@ public class ActivationController {
         final PrepareActivationRequest req = request.getRequestObject();
         logger.info("action: prepareActivation, state: initiated, applicationKey: {}", req.getApplicationKey());
         logger.debug("action: prepareActivation, state: initiated, request: {}", request);
-        final ObjectResponse<PrepareActivationResponse> response = new ObjectResponse<>(activationServiceBehaviorV3.prepareActivation(req));
+        final ObjectResponse<PrepareActivationResponse> response = new ObjectResponse<>(activationCreateServiceBehavior.prepareActivation(req));
         logger.info("action: prepareActivation, state: succeeded");
         logger.debug("action: prepareActivation, state: succeeded, response: {}", response);
         return response;
@@ -105,7 +109,7 @@ public class ActivationController {
         final CreateActivationRequest req = request.getRequestObject();
         logger.info("action: createActivation, state: initiated, userId: {}, applicationKey: {}", req.getUserId(), req.getApplicationKey());
         logger.debug("action: createActivation, state: initiated, request: {}", request);
-        final ObjectResponse<CreateActivationResponse> response = new ObjectResponse<>(activationServiceBehaviorV3.createActivation(req));
+        final ObjectResponse<CreateActivationResponse> response = new ObjectResponse<>(activationCreateServiceBehavior.createActivation(req));
         logger.info("action: createActivation, state: succeeded");
         logger.debug("action: createActivation, state: succeeded, response: {}", response);
         return response;
@@ -159,7 +163,7 @@ public class ActivationController {
         final GetActivationStatusRequest req = request.getRequestObject();
         logger.info("action: getActivationStatus, state: initiated, activationId: {}", req.getActivationId());
         logger.debug("action: getActivationStatus, state: initiated, request: {}", request);
-        final ObjectResponse<GetActivationStatusResponse> response = new ObjectResponse<>(activationServiceBehavior.getActivationStatus(req));
+        final ObjectResponse<GetActivationStatusResponse> response = new ObjectResponse<>(activationStatusServiceBehavior.getActivationStatus(req));
         logger.info("action: getActivationStatus, state: succeeded");
         logger.debug("action: getActivationStatus, state: succeeded, response: {}", response);
         return response;
